+++ R/load_semester_db.R
@@ -202,6 +202,11 @@ load_semester_db <- function(db_file, root_crit = NULL, ignore_root = FALSE) {
       dplyr::mutate(hw_num = seq(dplyr::n())) %>%
       dplyr::select(-"hw_grp_order")
 
+    if (!has_name(homework_items, "hw_self_assess")) {
+      warning("Database table homework_items is missing column hw_self_assess.",
+              "Setting to default of FALSE.")
+      homework_items <- homework_items %>% mutate(hw_self_assess = FALSE)
+    }
     hw_items <- homework_items %>%
       dplyr::inner_join(dplyr::select(homework_groups, "hw_grp_id", "hw_grp_key"),
                         by = "hw_grp_key") %>%
@@ -209,6 +214,7 @@ load_semester_db <- function(db_file, root_crit = NULL, ignore_root = FALSE) {
       dplyr::left_join(dplyr::select(hw_asgt, "hw_grp_id", "hw_num"),
                        by = "hw_grp_id") %>%
       dplyr::mutate(dplyr::across(c("undergraduate_only", "graduate_only",
+                                    "hw_self_assess",
                                     "hw_break_before", "hw_prologue", "hw_epilogue"),
                                   ~as.logical(.x) %>% tidyr::replace_na(FALSE)))
     if (!is.null(homework_solutions)) {